has phenotype and wraps around in space
