+++ Soup.py
@@ -1,19 +1,5 @@
 import numpy as np
-
-"""
-A Particle takes in 
-"""
-class Particle:
-    def __init__(self, x, y):
-        self.pos = np.array([x,y])
-        self.v = np.array([0,0]) #just to initialize
-        
-        
-    def move(self):
-        #This function needs to make an update based on the X = n + g + b rules
-        # TODO Right now I just want it to move left
-        self.v = np.random.normal(0, 1, (2,)) #making it a saved value so that it could be animated
-        self.pos += self.v
+from Particle import Particle
         
         
 class Soup:
@@ -21,8 +7,11 @@ class Soup:
     # defined width, heighth and number of randomlme generated particles
     def __init__(self, w, h, n):
         self.particles = []
+        self.w = w
+        self.h = h
+        pheno = dict(bias = np.array([1,2]))
         for _ in range(n):
-            self.particles.append(Particle(np.random.rand()*w, np.random.rand()*h))
+            self.particles.append(Particle(np.random.rand()*w, np.random.rand()*h, self, pheno))
             
     def move_all(self):
         for p in self.particles:
